move typeColors object to assets relates #35
