--- client/src/assets/index.js
@@ -1,4 +1,0 @@
-export { ReactComponent as Electricity } from './typesIcons/electricity.svg';
-export { ReactComponent as Wifi } from './typesIcons/wifi.svg';
-export { ReactComponent as Water } from './typesIcons/water.svg';
-export { ReactComponent as Mobile } from './typesIcons/smartphone-iphone.svg';
